Migrate invite_2b_claimer tests
