+++ libparsec/crates/protocol/tests/invite.rs
@@ -886,83 +886,6 @@ fn serde_invite_2b_greeter_send_nonce_rep(
 }
 
 #[parsec_test]
-fn serde_invite_2b_claimer_send_nonce_req() {
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   claimer_nonce: hex!("666f6f626172")
-    //   cmd: "invite_2b_claimer_send_nonce"
-    let raw = hex!(
-        "82ad636c61696d65725f6e6f6e6365c406666f6f626172a3636d64bc696e766974655f3262"
-        "5f636c61696d65725f73656e645f6e6f6e6365"
-    );
-
-    let req = invited_cmds::invite_2b_claimer_send_nonce::Req {
-        claimer_nonce: b"foobar".as_ref().into(),
-    };
-
-    let expected = invited_cmds::AnyCmdReq::Invite2bClaimerSendNonce(req);
-
-    let data = invited_cmds::AnyCmdReq::load(&raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = if let invited_cmds::AnyCmdReq::Invite2bClaimerSendNonce(data) = data {
-        data.dump().unwrap()
-    } else {
-        unreachable!()
-    };
-
-    let data2 = invited_cmds::AnyCmdReq::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
-
-#[parsec_test]
-#[case::ok(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "ok"
-    &hex!(
-        "81a6737461747573a26f6b"
-    )[..],
-    invited_cmds::invite_2b_claimer_send_nonce::Rep::Ok
-)]
-#[case::not_found(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "not_found"
-    &hex!(
-        "81a6737461747573a96e6f745f666f756e64"
-    )[..],
-    invited_cmds::invite_2b_claimer_send_nonce::Rep::NotFound
-)]
-#[case::invalid_state(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "invalid_state"
-    &hex!(
-        "81a6737461747573ad696e76616c69645f7374617465"
-    )[..],
-    invited_cmds::invite_2b_claimer_send_nonce::Rep::InvalidState
-)]
-fn serde_invite_2b_claimer_send_nonce_rep(
-    #[case] raw: &[u8],
-    #[case] expected: invited_cmds::invite_2b_claimer_send_nonce::Rep,
-) {
-    let data = invited_cmds::invite_2b_claimer_send_nonce::Rep::load(raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = data.dump().unwrap();
-
-    let data2 = invited_cmds::invite_2b_claimer_send_nonce::Rep::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
-
-#[parsec_test]
 fn serde_invite_3a_greeter_wait_peer_trust_req() {
     // Generated from Python implementation (Parsec v2.6.0+dev)
     // Content:
